+++ tests/test_components/test_heat_charge.py
@@ -2382,3 +2382,13 @@ def test_generation_recombination():
         E_0=1,
         sigma=2,
     )
+
+    # make sure we can build a SelberherrImpactIonization
+    _ = td.SelberherrImpactIonization(
+        alpha_n_inf=7.03e5,
+        alpha_p_inf=1.582e6,
+        E_n_crit=1.23e6,
+        E_p_crit=2.03e6,
+        beta_n=1,
+        beta_p=1,
+    )